+++ src/main/java/com/dnd5/timoapi/domain/user/presentation/response/UserServiceFeedbackDetailResponse.java
@@ -1,19 +1,27 @@
 package com.dnd5.timoapi.domain.user.presentation.response;
 
+import com.dnd5.timoapi.domain.test.presentation.response.TestResultResponse;
 import com.dnd5.timoapi.domain.user.domain.model.UserServiceFeedback;
+import com.dnd5.timoapi.domain.user.domain.model.UserTestRecord;
 
-public record UserServiceFeedbackResponse(
+import java.time.LocalDateTime;
+
+public record UserServiceFeedbackDetailResponse(
         Long id,
         Long userId,
         Long serviceRating,
-        String serviceFeedback
+        String serviceFeedback,
+        LocalDateTime createdAt
 ) {
-    public static UserServiceFeedbackResponse from(UserServiceFeedback model) {
-        return new UserServiceFeedbackResponse(
+    public static UserServiceFeedbackDetailResponse of(
+            UserServiceFeedback model
+    ) {
+        return new UserServiceFeedbackDetailResponse(
                 model.id(),
                 model.userId(),
                 model.serviceRating(),
-                model.serviceFeedback()
+                model.serviceFeedback(),
+                model.createdAt()
         );
     }
 }
